+++ Package.swift
@@ -43,6 +43,7 @@ let package = Package(
         .package(url: "https://github.com/SRGSSR/GoogleCastSDK-no-bluetooth.git", .upToNextMinor(from: "4.7.1-beta.1")),
         .package(url: "https://github.com/SRGSSR/TCCore-xcframework-apple.git", .upToNextMinor(from: "4.5.4-srg5")),
         .package(url: "https://github.com/SRGSSR/TCSDK-xcframework-apple.git", .upToNextMinor(from: "4.4.1-srg5")),
+        .package(url: "https://github.com/krzysztofzablocki/Difference.git", .upToNextMajor(from: "1.0.1")),
         .package(url: "https://github.com/Quick/Nimble.git", .upToNextMajor(from: "10.0.0"))
     ],
     targets: [
@@ -62,7 +63,13 @@ let package = Package(
                 .target(name: "Diagnostics")
             ]
         ),
-        .target(name: "Circumspect"),
+        .target(
+            name: "Circumspect",
+            dependencies: [
+                .product(name: "Difference", package: "Difference"),
+                .product(name: "Nimble", package: "Nimble")
+            ]
+        ),
         .target(name: "Diagnostics"),
         .target(name: "Player"),
         .target(
@@ -77,35 +84,34 @@ let package = Package(
             name: "AnalyticsTests",
             dependencies: [
                 .target(name: "Analytics"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect")
             ]
         ),
         .testTarget(
             name: "AppearanceTests",
             dependencies: [
                 .target(name: "Appearance"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect")
             ]
         ),
         .testTarget(
             name: "CircumspectTests",
             dependencies: [
-                .target(name: "Circumspect"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect")
             ]
         ),
         .testTarget(
             name: "CoreBusinessTests",
             dependencies: [
-                .target(name: "CoreBusiness"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect"),
+                .target(name: "CoreBusiness")
             ]
         ),
         .testTarget(
             name: "DiagnosticsTests",
             dependencies: [
-                .target(name: "Diagnostics"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect"),
+                .target(name: "Diagnostics")
             ]
         ),
         .testTarget(
@@ -113,14 +119,13 @@ let package = Package(
             dependencies: [
                 .target(name: "Circumspect"),
                 .target(name: "Player"),
-                .product(name: "Nimble", package: "Nimble")
             ]
         ),
         .testTarget(
             name: "UserInterfaceTests",
             dependencies: [
-                .target(name: "UserInterface"),
-                .product(name: "Nimble", package: "Nimble")
+                .target(name: "Circumspect"),
+                .target(name: "UserInterface")
             ]
         )
     ]
