+++ Tests/PlayerTests/Similarity.swift
@@ -6,9 +6,7 @@
 
 @testable import Player
 
-import Circumspect
 import CoreMedia
-import Nimble
 
 func beClose(within tolerance: TimeInterval) -> ((CMTime, CMTime) -> Bool) {
     Time.close(within: tolerance)
@@ -16,21 +14,4 @@ func beClose(within tolerance: TimeInterval) -> ((CMTime, CMTime) -> Bool) {
 
 func beClose(within tolerance: TimeInterval) -> ((CMTimeRange, CMTimeRange) -> Bool) {
     Time.close(within: tolerance)
-}
-
-/// Match against an expected value using some comparator.
-/// Directly borrowed from https://github.com/Quick/Nimble/blob/main/Sources/Nimble/Matchers/Equal.swift
-func equal<T>(_ expectedValue: T?, by areEquivalent: @escaping (T, T) -> Bool) -> Predicate<T> {
-    Predicate.define("equal <\(stringify(expectedValue))>") { actualExpression, message in
-        let actualValue = try actualExpression.evaluate()
-        switch (expectedValue, actualValue) {
-        case (nil, _?):
-            return PredicateResult(status: .fail, message: message.appendedBeNilHint())
-        case (_, nil):
-            return PredicateResult(status: .fail, message: message)
-        case let (expected?, actual?):
-            let matches = areEquivalent(expected, actual)
-            return PredicateResult(bool: matches, message: message)
-        }
-    }
 }
