+++ Tests/PlayerTests/TimeTests.swift
@@ -7,6 +7,7 @@
 @testable import Player
 
 import AVFoundation
+import Circumspect
 import Nimble
 import XCTest
